+++ bulk_data_gen/common/distribution.go
@@ -1,4 +1,4 @@
-package main
+package common
 
 import (
 	"math"
