+++ bulk_data_gen/common/point.go
@@ -1,4 +1,4 @@
-package main
+package common
 
 import (
 	"encoding/binary"
