+++ bulk_data_gen/common/pools.go
@@ -1,4 +1,4 @@
-package main
+package common
 
 import (
 	"sync"
