+++ bulk_data_gen/common/simulation.go
@@ -1,4 +1,4 @@
-package main
+package common
 
 import "time"
