+++ bulk_data_gen/devops/devops_host.go
@@ -1,8 +1,9 @@
-package main
+package devops
 import (
 	"fmt"
 	"math/rand"
 	"time"
+	. "github.com/influxdata/influxdb-comparisons/bulk_data_gen/common"
 )
 
 const NHostSims = 9
