+++ bulk_data_gen/devops/devops_mem.go
@@ -1,9 +1,10 @@
-package main
+package devops
 
 import (
 	"math"
 	"math/rand"
 	"time"
+	. "github.com/influxdata/influxdb-comparisons/bulk_data_gen/common"
 )
 
 var (
